% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_bad_argument}
\alias{abort_bad_argument}
\title{Utilities}
\usage{
abort_bad_argument(arg, must, not = NULL)
}
\arguments{
\item{arg}{the argument}

\item{must}{allowed values of argument}

\item{not}{the bare argument}
}
\value{
utilities
}
\description{
abort_bad_argument taken from Advanced R! Written by Hadley Wichkham.
}
