% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateview_plot.R
\name{add_empty_wells}
\alias{add_empty_wells}
\title{Add empty wells to a partially-filled layout}
\usage{
add_empty_wells(
  df,
  n_wells = "384",
  .df_well_col = "well",
  .fill_down_cols = "",
  add_rows_cols = TRUE
)
}
\arguments{
\item{df}{the layout file to have empty wells appended}

\item{n_wells}{the number of wells in the plate. Options are "384" and "96". Defaults to "384".}

\item{.df_well_col}{the name of the colum containing well information. Defaults to "well".}

\item{.fill_down_cols}{a vector of any columns which should be filled in for the empty wells. These will be filled with a single value, determined by the plyr::fill(.direction = "down") function.}

\item{add_rows_cols}{should plate rows and columns be added to the layout (e.g. A, B, . . ., 1, 2, 3, ...)? Defaults to TRUE.}
}
\value{
the input layout, with all empty wells appended. These values are NA, unless specified in the .fill_down_cols argument of this function.
}
\description{
A helper function for making plate plots using the plate_plot() function
}
