% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateview_plot.R
\name{all_plateview_vars}
\alias{all_plateview_vars}
\title{Create a single tibble summarizing the transfer conditions and any important changes and repairs}
\usage{
all_plateview_vars(daughter_raw, transfers, depletion)
}
\arguments{
\item{daughter_raw}{the standardized daughter}

\item{transfers}{the transfers tibble}

\item{depletion}{the depletion tibble}
}
\value{
a list, with elements data: a tibble containins pertinent columns from the three input tibbles, and plot_vars, a list of variables from this tibble to be passed to make_all_plots
}
\description{
Create a single tibble summarizing the transfer conditions and any important changes and repairs
}
