% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_updated_layout.R
\name{bind_layouts}
\alias{bind_layouts}
\title{Make wideform layouts for multiple variables}
\usage{
bind_layouts(data, var_list)
}
\arguments{
\item{data}{a layout}

\item{var_list}{the columns in the layout to be make into individual layout blocks}
}
\value{
a tibble of wideform layout blocks, ready to be saved as a csv
}
\description{
A helper function for make_updated_layout().
Relies on its own helper function, make_layout_wide().
Ultimately, this function and its helpers should be a part of a different package--likely the layouts package, if we ever get that thing done.
}
