% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transfers.R
\name{calculate_transfers}
\alias{calculate_transfers}
\title{Calculate transfer steps from layouts}
\usage{
calculate_transfers(
  daughter,
  mother,
  .echo_drop_nL = 25,
  .dilutant_name = "DMSO"
)
}
\arguments{
\item{daughter}{A daughter layout, standardized and repaired}

\item{mother}{A mother layout, standardized and repaired}

\item{.echo_drop_nL}{The transfer volume of the Echo instrument to be used, in nL. Defaults to 25.}

\item{.dilutant_name}{The name of the compound in the mother to be used for dilutions. Defaults to "DMSO".}
}
\value{
A tibble containing all transfer steps, and final conditions representing any necessary rounding
}
\description{
Calculate transfer steps from layouts
}
