% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transfers.R
\name{concentrations_to_transfers}
\alias{concentrations_to_transfers}
\title{Concentrations to transfers}
\usage{
concentrations_to_transfers(daughter, mother, .echo_drop_nL = 25)
}
\arguments{
\item{daughter}{repaired daughter layout}

\item{mother}{repaired mother layout}

\item{.echo_drop_nL}{the droplet size for the echo, in nL. Defaults to 25.}
}
\value{
tibble of calcualted transfers, withoutdistribution across source
}
\description{
Concentrations to transfers
}
