% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transfers.R
\name{distribute_shared}
\alias{distribute_shared}
\title{Distribute transfers over multiple mother source wells}
\usage{
distribute_shared(transfers, .echo_drop_nL = 25)
}
\arguments{
\item{transfers}{a tibble of transfers}

\item{.echo_drop_nL}{droplet size of echo in nL. Defaults to 25.}
}
\value{
a modified tibble of transfers, with transfers distributed across common mother wells.
}
\description{
Distribute transfers over multiple mother source wells
}
