% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateview_plot.R
\name{get_plotworthy_vars}
\alias{get_plotworthy_vars}
\title{Extract variables worth making plateview plots}
\usage{
get_plotworthy_vars(
  data,
  drop_vars = c("Destination Well", "Source Well", "well", "Well", "row", "column",
    "mother_vol", "rounded_up_perc", "mother_conc")
)
}
\arguments{
\item{data}{layout tibble}

\item{drop_vars}{vector of variables not worth plotting. Defaults to: c("Destination Well","Source Well","well","Well","row","column","mother_vol","rounded_up_perc","mother_conc" )}
}
\value{
a character vector of variables from the input data maybe work making into a plateview plot.
}
\description{
Mostly just here to mask out things like wells, rows, and columns
}
