% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts_to_instructions.R
\name{layouts_to_instructions}
\alias{layouts_to_instructions}
\title{The one-shot function to convert layouts into an echo instructions file, and save the associated fodler}
\usage{
layouts_to_instructions(
  expnum,
  save_path = "",
  daughter_path,
  mother_path,
  save_output = FALSE,
  .well_col_daughter = "well",
  .compound_col_daughter = "compound",
  .concentration_col_daughter = "concentration",
  .volume_col_daughter = "volume",
  .well_col_mother = "well",
  .compound_col_mother = "compound",
  .concentration_col_mother = "concentration",
  if_missing = "stop",
  if_varied = "stop",
  if_impossible = "stop",
  echo_drop_vol = 25,
  max_mother_pull = 35
)
}
\arguments{
\item{expnum}{The experiment number, e.g. "4242"}

\item{save_path}{the path to save the instructions folder in}

\item{daughter_path}{path to the daughter layout}

\item{mother_path}{to the mother layout}

\item{save_output}{if TRUE, saves the resulting instructions in the save_path}

\item{.well_col_daughter}{passed to standardize_layouts for daughter. Defaults to "well".}

\item{.compound_col_daughter}{passed to standardize_layouts for daughter. Defaults to "compound".}

\item{.concentration_col_daughter}{passed to standardize_layouts for daughter. Defaults to "concentration".}

\item{.volume_col_daughter}{passed to standardize_layouts for daughter. Defaults to "volume".}

\item{.well_col_mother}{passed to standardize_layouts for mother. Defaults to "well".}

\item{.compound_col_mother}{passed to standardize_layouts for mother. Defaults to "compound".}

\item{.concentration_col_mother}{passed to standardize_layouts for mother. Defaults to "concentration".}

\item{if_missing}{action to take if compounds in the daughter are not in the mother. Defaults to "stop". Passed to repair_layout.}

\item{if_varied}{action to take if ccompounds are present at multile concentrations in the mother. Defaults to "stop". Passed to repair_layout.}

\item{if_impossible}{action to take if compounds are present in the daughter at concentrations greater than the mother can provide. Defaults to "stop". Passed to repair_layout.}

\item{echo_drop_vol}{the volume of the droplet for the echo instrument to be used. Defaults to 25.}

\item{max_mother_pull}{the maximum volume to be pulled from any one well in the mother. If exceeded, this will be noted in the the final plateview plots, and transfer csv. Defaults to 35.}
}
\value{
a list containing all intermediate and final objects. If save_output = TRUE, also saves the echo instrutions in a folder.
}
\description{
The one-shot function to convert layouts into an echo instructions file, and save the associated fodler
}
