% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateview_plot.R
\name{make_all_plots}
\alias{make_all_plots}
\title{make_all_plots}
\usage{
make_all_plots(plate_data, plot_these)
}
\arguments{
\item{plate_data}{the layout tibble from which the plots will be made}

\item{plot_these}{a list of column names in plate_data--a plateview plot will be made for each element in this list.}
}
\value{
A list, containing (1) "individual_plots", which has each individual variable plotted, stored as a sub-itme under the name of the variable it contains. (2) "all_plots_fig", a ggpubr object containing all of the plotted variables in a single figure, useful for simple batch download. (3) save_width: the recommended widith to save the ggpubr object, set to 12. (4) save_height: the recommended height to save the all_plots_fig, calculated to depend on the number of plots that this figure contains.
}
\description{
Make a plateview plot for a every variable in a provided list.
}
