% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transfers.R
\name{make_dilutions_plate}
\alias{make_dilutions_plate}
\title{Calculate dilution trasnfers}
\usage{
make_dilutions_plate(
  transfers,
  mother,
  .echo_drop_nL = 25,
  .dilutant_name = "DMSO"
)
}
\arguments{
\item{transfers}{tibble containing compound transferd}

\item{mother}{repaired mother layout}

\item{.echo_drop_nL}{droplet size of echo in nL. Defaults to 25.}

\item{.dilutant_name}{name of the compound used for diluations. Defaults to "DMSO"}
}
\value{
a transfer tibble, specifically for the dilutions.
}
\description{
Calculate dilution trasnfers
}
