% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_exp_header}
\alias{make_exp_header}
\title{Make experiment header}
\usage{
make_exp_header(expnum = "0000")
}
\arguments{
\item{expnum}{the experiment number. Defaults to "0000"}
}
\value{
a string fo the standard experiment number prefix format: "Exp<expnum>--YYYYMMDD_"
}
\description{
Not sure what package this really belong in in the end, but its useful here, so adding it here for now.
}
