% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_updated_layout.R
\name{make_layout_wide}
\alias{make_layout_wide}
\title{Convert a single layout variable into a wideform layout block}
\usage{
make_layout_wide(data, .filt_col, .fill_empties = "Empty")
}
\arguments{
\item{data}{the layout}

\item{.filt_col}{a single column in data to be converted into a layout block}

\item{.fill_empties}{what to fill in for wells without contents. Defaults to "Empty"}
}
\value{
A tibble, containing a wideform layout block.
}
\description{
A helper function to make_updated_layout.
Ultimately, this function should be a part of a different package--likely the layouts package, if we ever get that thing done.
}
