% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_updated_layout.R
\name{make_updated_layout}
\alias{make_updated_layout}
\title{Update starting layout with final conditions}
\usage{
make_updated_layout(transfers, raw_layout)
}
\arguments{
\item{transfers}{the transfers tibble}

\item{raw_layout}{the raw layout, as uploaded by the user}
}
\value{
a list of 2 elements. layout, the updated layout, and wide_layout, the updated layout in wideform, ready to be saved as a csv.
}
\description{
Update starting layout with final conditions
}
