% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transfers.R
\name{monitor_source_depletion}
\alias{monitor_source_depletion}
\title{Monitor depletion of source plate wells}
\usage{
monitor_source_depletion(transfers, max_uL_pull = 35)
}
\arguments{
\item{transfers}{A tibble containing the final calculated transfers.}

\item{max_uL_pull}{The maximum volume that can be pulled from any single mother well. Defaults to 35.}
}
\value{
A tibble containing Source Wells, compounds, and total volume pulled from each well
}
\description{
Monitor depletion of source plate wells
}
