% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_layout.R
\name{repair_layout}
\alias{repair_layout}
\title{Repair impossible transfer requests}
\usage{
repair_layout(
  mother,
  daughter,
  if_missing = "stop",
  if_varied = "stop",
  if_impossible = "stop"
)
}
\arguments{
\item{mother}{A standardized mother layout.}

\item{daughter}{A standardized daughter layout.}

\item{if_missing}{How to resolve cases where a compound is present in the daughter, but not the mother.
\itemize{
\item{"stop": Halt execution, and return an informative error message.}
\item{"drop": Remove these compounds from the daughter layout.}
}}

\item{if_varied}{How to resolve cases where a single compound is present in the mother at multiple concentrations.
\itemize{
\item{"stop": Halt execution, and return an informative error message.}
\item{"keep_max": Transfer only from wells containing the highest concentration of this compound.}
\item{"keep_most": Transfer only from wells containing the most abundant concentration of this compound.}
}}

\item{if_impossible}{How to resolve cases where a compound in the daughter at a concentration exceeding its concentration in the mother.
\itemize{
\item{"stop": Halt execution, and return an informative error message.}
\item{"drop": Remove these concentrations from the daughter layout.}
\item{"make_max": Replace these concentrations with the highest concentration achievable with the given mother plate.}
\item{"scale_down": Replace all concentartions of offending compounds with a scaled-down concentration, such that the relative concentrations of this compound are retained, and all requested concentrations are equal to or less than the concentration of this compound in the mother plate. }
}}
}
\value{
A list
\itemize{
\item{mother: a mother plate, with any issues repaired as specified}
\item{daughter: a daughter plate, with any issues repaired as specified}}
}
\description{
Repairs mother and daughter layouts resulting impossible transfers. Allows users to choose options to minimally modify layouts to remove impossible transfers.
}
