% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_instructions_folder.R
\name{save_instructions_folder}
\alias{save_instructions_folder}
\title{Save echo instructions folder}
\usage{
save_instructions_folder(
  expnum,
  save_path = "",
  final_layouts,
  instructions,
  all_plots,
  layouts,
  transfers,
  depletion,
  version_num = "0.0.0.9000"
)
}
\arguments{
\item{expnum}{The experiment number, including wordy details, e.g "4242_daugher_plates"}

\item{save_path}{The path to where the folder should be saved. Will proceed the experiment number and "_echo instructions". Defaults to "", saving the folder as "Exp<expnum>_echo_instructions/"}

\item{final_layouts}{The final layouts list, created by make_updated_layout()}

\item{instructions}{The echo instructions tibble, created by write_instructions_file()}

\item{all_plots}{The plateview plots, created by make_all_plots()}

\item{layouts}{The standardized layouts, created by repair_layout()}

\item{transfers}{The calculated transfers, created by calculate_transfers()}

\item{depletion}{The tibble of mother plate depletions, created by monitor_source_depletion()}

\item{version_num}{The version number for the echowrtir package. Gets written into the reademe; is actually the only thing in the readme.txt at this point...}
}
\value{
A folder in the given directory, containing everything needed to echo your plate, and accurately manage the resulting data, including any rounding or omissions.
}
\description{
Save echo instructions folder
}
