% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_layout.R
\name{standardize_layout}
\alias{standardize_layout}
\title{Standardize plate layouts}
\usage{
standardize_layout(
  layout,
  which_plate,
  .well_col = "well",
  .compound_col = "compound",
  .concentration_col = "concentration",
  .volume_col = "volume"
)
}
\arguments{
\item{layout}{A plate layout, containing columns with compound, concentration, and, for daughter plates only, volume.}

\item{which_plate}{The type of layout to standardize. Can be "mother" or "daughter".}

\item{.well_col, .compound_col, .concentration_col, .volume_col}{Names of the columns in the input layout containg necessary information.
\itemize{
\item{.well_col: The well names.}
\item{.compound_col: The compounds to be transferred. Compounds must go by the same name in mother and daughter plates.}
\item{.concentration_col: The concentration of each compound, either in the mother plate, or the desired concentration in the daughter plate.}
\item{.volume_vol: For daughter plates only. The desired volume in nL in each well of the daughter. }
}}
}
\value{
A plate layout, containing standardized columns and column names for use in downstream echowritr functions.
}
\description{
Standardize plate layouts
}
