% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_rounding.R
\name{summarise_rounding}
\alias{summarise_rounding}
\title{Summarise the rounding resulting from the transfer calculations}
\usage{
summarise_rounding(transfers, return_filtered = FALSE)
}
\arguments{
\item{transfers}{the transfers tibble}

\item{return_filtered}{if TRUE, returns a tibble containins only rows corresponding to rounded values. If FALSE, no filtering is performed.}
}
\value{
the transfers tibbl, with only the columns relevant to rounding. These columns are renamed so that users can guess at their meaning. Verbose but hopefully clearer.
}
\description{
Wil likely get built out further for a more comprehensive changes report.
}
