% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateview_plot.R
\name{plateview_plot}
\alias{plateview_plot}
\title{Make a Plate-view plot}
\usage{
plateview_plot(
  layout_data,
  .fill_var,
  .well_col = "well",
  col_breaks = 6,
  .title = "Plate-view plot",
  .scale_fill = scale_fill_viridis_d()
)
}
\arguments{
\item{layout_data}{a layout tibble (as created by dsfworld::read_plate_layout()), containing columns for plate row (called "row"), plate column (called "column"), and one variable by which the plate will be colored}

\item{.fill_var}{the column in the input layout by which to color in the wells in the plot}

\item{.well_col}{the name of the column containing the wells. Defaults to "well".}

\item{col_breaks}{the number of tick marks and lines in the x axis (columns). Defaults to every 6 columns.}

\item{.title}{title for the plot. Defaults to "Plate-view plot"}

\item{.scale_fill}{the color scale to use in the plot. Defaults to scale_color_viridis_d(). Must be updated when plotting continuous variables (e.g. to scale_color_viridis_c())}
}
\value{
a plate-view plot, with wells colored based on the information on that well in the user-defined variable given in .fill_var.
}
\description{
Makes a plot that mimics looking down at a well plate. Wells are colored based on a user-defined column from a layout.
}
